Add ISerializeFormatter
