+++ Runtime/Core/Serialization/SaveUtility.cs
@@ -1,143 +1,8 @@
-using System;
 using System.IO;
-using System.Reflection;
-using System.Runtime.Serialization.Formatters.Binary;
-using Newtonsoft.Json;
 using UnityEngine;
 
 namespace Chris.Serialization
 {
-    /// <summary>
-    /// Serializer for read and write saving files
-    /// </summary>
-    public class SaveLoadSerializer
-    {
-        public static class TypeCache<T>
-        {
-            public static readonly bool PreferJsonConvert;
-
-            static TypeCache()
-            {
-                PreferJsonConvert = typeof(T).GetCustomAttribute<PreferJsonConvertAttribute>() != null;
-            }
-        }
-        
-        private static readonly BinaryFormatter Formatter = new();
-
-        private readonly string _path;
-        
-        private readonly string _extension;
-
-        private const string DefaultExtension = "sav";
-
-        public SaveLoadSerializer(string path, string extension = DefaultExtension)
-        {
-            _path = path;
-            _extension = extension;
-        }
-        
-        public void Save<T>(string key, T data)
-        {
-            string jsonData;
-            if (TypeCache<T>.PreferJsonConvert)
-                jsonData = JsonConvert.SerializeObject(data);
-            else
-                jsonData = JsonUtility.ToJson(data);
-            if (!Directory.Exists(_path)) Directory.CreateDirectory(_path);
-            using var file = File.Create($"{_path}/{key}.{_extension}");
-            Formatter.Serialize(file, jsonData);
-        }
-        
-        public void Delete(string key)
-        {
-            if (!Directory.Exists(_path)) return;
-            var path = $"{_path}/{key}.{_extension}";
-            if (File.Exists(path))
-            {
-                File.Delete(path);
-            }
-        }
-        
-        public void DeleteAll()
-        {
-            if (Directory.Exists(_path)) Directory.Delete(_path, true);
-        }
-        
-        public void SaveJson(string key, string jsonData)
-        {
-            if (!Directory.Exists(_path)) Directory.CreateDirectory(_path);
-            using var file = File.Create($"{_path}/{key}.{_extension}");
-            Formatter.Serialize(file, jsonData);
-        }
-        
-        public bool Exists(string key)
-        {
-            return File.Exists($"{_path}/{key}.{_extension}");
-        }
-        
-        public bool TryLoadJson(string key, out string jsonData)
-        {
-            var path = $"{_path}/{key}.{_extension}";
-            if (File.Exists(path))
-            {
-                using var file = File.Open(path, FileMode.Open);
-                jsonData = (string)Formatter.Deserialize(file);
-                return true;
-            }
-            jsonData = null;
-            return false;
-        }
-
-        public bool Overwrite<T>(string key, T data)
-        {
-            var path = $"{_path}/{key}.{_extension}";
-            if (File.Exists(path))
-            {
-                using var file = File.Open(path, FileMode.Open);
-                if (TypeCache<T>.PreferJsonConvert)
-                    JsonConvert.PopulateObject((string)Formatter.Deserialize(file), data);
-                else
-                    JsonUtility.FromJsonOverwrite((string)Formatter.Deserialize(file), data);
-                return true;
-            }
-            return false;
-        }
-        
-        public T LoadOrNew<T>(string key) where T : class, new()
-        {
-            T data = null;
-            var path = $"{_path}/{key}.{_extension}";
-            if (File.Exists(path))
-            {
-                using var file = File.Open(path, FileMode.Open);
-                if (TypeCache<T>.PreferJsonConvert)
-                    data = JsonConvert.DeserializeObject<T>((string)Formatter.Deserialize(file));
-                else
-                    data = JsonUtility.FromJson<T>((string)Formatter.Deserialize(file));
-            }
-            data ??= new T();
-            return data;
-        }
-        
-        public T Load<T>(string key)
-        {
-            var path = $"{_path}/{key}.{_extension}";
-            using var file = File.Open(path, FileMode.Open);
-            if (TypeCache<T>.PreferJsonConvert)
-                return JsonConvert.DeserializeObject<T>((string)Formatter.Deserialize(file));
-            return JsonUtility.FromJson<T>((string)Formatter.Deserialize(file));
-        }
-        
-        public object Load(string key, Type type, bool preferJsonConvert)
-        {
-            var path = $"{_path}/{key}.{_extension}";
-            using var file = File.Open(path, FileMode.Open);
-            if (preferJsonConvert)
-                return JsonConvert.DeserializeObject((string)Formatter.Deserialize(file), type);
-            return JsonUtility.FromJson((string)Formatter.Deserialize(file), type);
-        }
-    }
-    
     public static class SaveUtility
     {
         static SaveUtility()
